Move drakeGradientUtil.h from util/ to core/. It is used by files in core, so cannot be placed later in the build. Leave util/drakeGraidientUtil.h around as a convenient forwarder.
